Fixing more bugs
I fixed the error in no_level where there's no image showing and the link always going to nowhere. I also fixed a bug where the database error shows up whenever I'm processing the non-existent ID (this applies to a lot of id specific inputs, but not all yet). Lastly, I also fixed a bug where the footer is being horizontal. I also moved all of session managing from views to controller, where I can see everything there, so the views only used for showing data only.

+++ application/controllers/C_tabel_b1.php
@@ -15,6 +15,7 @@ class C_tabel_b1 extends Omnitags
 	public function admin()
 	{
 		$this->declarew();
+		$this->page_session_3();
 
 		$param1 = $this->v_get['tabel_b1_field7'];
 
@@ -37,6 +38,7 @@ class C_tabel_b1 extends Omnitags
 	public function filter()
 	{
 		$this->declarew();
+		$this->page_session_3();
 
 		validate_input(
 			array(
@@ -138,70 +140,83 @@ class C_tabel_b1 extends Omnitags
 
 		$tabel_b1_field1 = $this->v_post['tabel_b1_field1'];
 
-		validate_input(
-			array(
-				$this->v_post['tabel_b1_field1'],
-				$this->v_post['tabel_b1_field2'],
-				$this->v_post['tabel_b1_field3'],
-				$this->v_post['tabel_b1_field5'],
-				$this->v_post['tabel_b1_field6'],
-				$this->v_post['tabel_b1_field7'],
-			),
-			$this->views['flash3'],
-			'ubah' . $tabel_b1_field1
-		);
+		$tabel = $this->tl_b1->get_b1_by_b1_field1($tabel_b1_field1)->result();
 
-		$tabel_b1 = $this->tl_b1->get_b1_by_b1_field1($tabel_b1_field1)->result();
-		$new_name = $this->v_post['tabel_b1_field2'];
-		$path = $this->v_upload_path['tabel_b1'];
-		$img = $this->v_post['tabel_b1_field4_old'];
-		$extension = '.' . getExtension($path . $img);
+		if ($tabel) {
 
-		$config['upload_path'] = $path;
-		// nama file telah ditetapkan dan hanya berekstensi jpg dan dapat diganti dengan file bernama sama
-		$config['file_name'] = $new_name;
-		$config['allowed_types'] = $this->file_type1;
-		$config['overwrite'] = TRUE;
-		$config['remove_spaces'] = TRUE;
+			validate_input(
+				array(
+					$this->v_post['tabel_b1_field1'],
+					$this->v_post['tabel_b1_field2'],
+					$this->v_post['tabel_b1_field3'],
+					$this->v_post['tabel_b1_field5'],
+					$this->v_post['tabel_b1_field6'],
+					$this->v_post['tabel_b1_field7'],
+				),
+				$this->views['flash3'],
+				'ubah' . $tabel_b1_field1
+			);
 
-		$this->load->library('upload', $config);
-		$upload = $this->upload->do_upload($this->v_input['tabel_b1_field4_input']);
+			$tabel_b1 = $this->tl_b1->get_b1_by_b1_field1($tabel_b1_field1)->result();
+			$new_name = $this->v_post['tabel_b1_field2'];
+			$path = $this->v_upload_path['tabel_b1'];
+			$img = $this->v_post['tabel_b1_field4_old'];
+			$extension = '.' . getExtension($path . $img);
 
-		if (!$upload) {
-			if ($new_name != $tabel_b1[0]->kode) {
-				rename($path . $img, $path . str_replace(' ', '_', $new_name) . $extension);
-				$gambar = str_replace(' ', '_', $new_name) . $extension;
-			} else {
-				$gambar = $img;
-			}
-		} else {
-			if ($new_name != $tabel_b1[0]->kode) {
-				// File upload is successful, delete the old file
-				if (file_exists($path . $img)) {
-					unlink($path . $img);
+			$config['upload_path'] = $path;
+			// nama file telah ditetapkan dan hanya berekstensi jpg dan dapat diganti dengan file bernama sama
+			$config['file_name'] = $new_name;
+			$config['allowed_types'] = $this->file_type1;
+			$config['overwrite'] = TRUE;
+			$config['remove_spaces'] = TRUE;
+
+			$this->load->library('upload', $config);
+			$upload = $this->upload->do_upload($this->v_input['tabel_b1_field4_input']);
+
+			if (!$upload) {
+				if ($new_name != $tabel_b1[0]->kode) {
+					rename($path . $img, $path . str_replace(' ', '_', $new_name) . $extension);
+					$gambar = str_replace(' ', '_', $new_name) . $extension;
+				} else {
+					$gambar = $img;
 				}
-				$upload = $this->upload->data();
-				$gambar = $upload['file_name'];
 			} else {
-				$gambar = $img;
+				if ($new_name != $tabel_b1[0]->kode) {
+					// File upload is successful, delete the old file
+					if (file_exists($path . $img)) {
+						unlink($path . $img);
+					}
+					$upload = $this->upload->data();
+					$gambar = $upload['file_name'];
+				} else {
+					$gambar = $img;
+				}
 			}
+
+			// menggunakan nama khusus sama dengan konfigurasi
+			$data = array(
+				$this->aliases['tabel_b1_field2'] => $this->v_post['tabel_b1_field2'],
+				$this->aliases['tabel_b1_field3'] => $this->v_post['tabel_b1_field3'],
+				$this->aliases['tabel_b1_field4'] => $gambar,
+				$this->aliases['tabel_b1_field5'] => $this->v_post['tabel_b1_field5'],
+				$this->aliases['tabel_b1_field6'] => $this->v_post['tabel_b1_field6'],
+				$this->aliases['tabel_b1_field7'] => $this->v_post['tabel_b1_field7'],
+			);
+
+			$aksi = $this->tl_b1->update_b1($data, $tabel_b1_field1);
+
+			$notif = $this->handle_4c($aksi, 'tabel_b1', $tabel_b1_field1);
+
+			redirect($_SERVER['HTTP_REFERER']);
+
+
+
+		} else {
+			// error handling
+			set_flashdata($this->views['flash1'], "Error occurred while processing data!");
+			set_flashdata('toast', $this->views['flash1_func1']);
+			redirect(userdata('previous_url'));
 		}
-
-		// menggunakan nama khusus sama dengan konfigurasi
-		$data = array(
-			$this->aliases['tabel_b1_field2'] => $this->v_post['tabel_b1_field2'],
-			$this->aliases['tabel_b1_field3'] => $this->v_post['tabel_b1_field3'],
-			$this->aliases['tabel_b1_field4'] => $gambar,
-			$this->aliases['tabel_b1_field5'] => $this->v_post['tabel_b1_field5'],
-			$this->aliases['tabel_b1_field6'] => $this->v_post['tabel_b1_field6'],
-			$this->aliases['tabel_b1_field7'] => $this->v_post['tabel_b1_field7'],
-		);
-
-		$aksi = $this->tl_b1->update_b1($data, $tabel_b1_field1);
-
-		$notif = $this->handle_4c($aksi, 'tabel_b1', $tabel_b1_field1);
-
-		redirect($_SERVER['HTTP_REFERER']);
 	}
 
 	public function sync_theme($tabel_b1_field7 = null)
@@ -209,15 +224,26 @@ class C_tabel_b1 extends Omnitags
 		$this->declarew();
 		$this->session_3();
 
-		$data = array(
-			$this->aliases['tabel_b1_field7'] => $tabel_b1_field7,
-		);
+		$tabel = $this->tl_b7->get_b7_by_b7_field1($tabel_b1_field7)->result();
 
-		$aksi = $this->tl_b1->update_all_b1($data);
+		if ($tabel) {
 
-		$notif = $this->handle_4c($aksi, 'tabel_b1', $tabel_b1_field7);
+			$data = array(
+				$this->aliases['tabel_b1_field7'] => $tabel_b1_field7,
+			);
 
-		redirect($_SERVER['HTTP_REFERER']);
+			$aksi = $this->tl_b1->update_all_b1($data);
+
+			$notif = $this->handle_4c($aksi, 'tabel_b1', $tabel_b1_field7);
+
+			redirect($_SERVER['HTTP_REFERER']);
+
+		} else {
+			// error handling
+			set_flashdata($this->views['flash1'], "Error occurred while processing data!");
+			set_flashdata('toast', $this->views['flash1_func1']);
+			redirect(userdata('previous_url'));
+		}
 	}
 
 	public function delete($tabel_b1_field1 = null)
@@ -226,21 +252,31 @@ class C_tabel_b1 extends Omnitags
 		$this->session_3();
 
 		$tabel_b1 = $this->tl_b1->get_b1_by_b1_field1($tabel_b1_field1)->result();
-		$img = $tabel_b1[0]->img;
 
-		unlink($this->v_upload_path['tabel_b1'] . $img);
+		if ($tabel_b1) {
+			$img = $tabel_b1[0]->img;
 
-		$aksi = $this->tl_b1->delete_b1($tabel_b1_field1);
+			unlink($this->v_upload_path['tabel_b1'] . $img);
 
-		$notif = $this->handle_4e($aksi, 'tabel_b1', $tabel_b1_field1);
+			$aksi = $this->tl_b1->delete_b1($tabel_b1_field1);
 
-		redirect($_SERVER['HTTP_REFERER']);
+			$notif = $this->handle_4e($aksi, 'tabel_b1', $tabel_b1_field1);
+
+			redirect($_SERVER['HTTP_REFERER']);
+
+		} else {
+			// error handling
+			set_flashdata($this->views['flash1'], "Error occurred while processing data!");
+			set_flashdata('toast', $this->views['flash1_func1']);
+			redirect(userdata('previous_url'));
+		}
 	}
 
 	// Cetak semua data
 	public function laporan()
 	{
 		$this->declarew();
+		$this->page_session_3();
 
 		$data1 = array(
 			'title' => lang('tabel_b1_alias_v4_title'),
